requirements.txt
absl-py==0.7.1
alembic==1.0.5
aniso8601==4.0.1
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
autopep8==1.4.4
Babel==2.6.0
beautifulsoup4==4.8.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.4.2
DateTime==4.3
defusedxml==0.5.0
Flask==1.0.2
Flask-AppBuilder==1.12.2
Flask-Babel==0.12.2
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-MySQLdb==0.2.0
Flask-OpenID==1.2.5
Flask-RESTful==0.3.7
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.17.1
gast==0.2.2
get==2018.11.19
google-pasta==0.1.7
googletrans==2.4.0
grpcio==1.22.0
h5py==2.9.0
haversine==2.0.0
hydra-framework==0.2.2
idna==2.7
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.0
line-bot-sdk==1.8.0
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.0
mccabe==0.6.1
mimerender==0.6.0
mysqlclient==1.3.14
Naked==0.1.31
nltk==3.4
numpy==1.17.0
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.1.0
post==2018.11.20
protobuf==3.9.1
psycopg2==2.7.7
public==2018.11.20
pycodestyle==2.5.0
pycparser==2.19
PyExecJS==1.5.1
pylint==2.3.1
pylint-flask==0.6
pylint-plugin-utils==0.5
PyMySQL==0.9.2
pyopengl==3.1.3b2
pyspellchecker==0.4.0
python-dateutil==2.7.5
python-editor==1.0.3
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2018.7
PyYAML==5.1
query-string==2018.11.20
request==2018.11.20
requests==2.20.1
Sastrawi==1.0.1
scipy==1.3.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.2.14
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
textblob==0.15.2
typed-ast==1.3.5
urllib3==1.24.1
virtualenv==16.4.0
Werkzeug==0.15.5
wrapt==1.11.2
WTForm==1.0
WTForms==2.2.1
zope.interface==4.6.0
